# api/routes/openapi/openapi.json
{
  "openapi": "3.1.0",
  "servers": [
    {
      "url": "/v1"
    }
  ],
  "info": {
    "title": "IndexCards API",
    "version": "1.2.0",
    "description": "Tabroom.com data & operational API",
    "termsOfService": "https://www.speechanddebate.org/terms-conditions/",
    "license": {
      "name": "Copyright 2014-2021, National Speech & Debate Association",
      "identifier": "RPL-1.5"
    }
  },
  "security": [
    {
      "bearer": []
    },
    {
      "cookie": []
    }
  ],
...
  "paths": {
...
    "/rest/tourns/{tournId}/events/{eventAbbr}": {
      "get": {
        "summary": "Returns some limited data about an event together with published rounds by event abbreviation",
        "operationId": "getEventByAbbr",
        "responses": {
          "200": {
            "description": "Event and Round in JSON format for parsing",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "tags": [
          "invite",
          "public",
          "event",
          "eventAbbr",
          "rounds"
        ],
        "description": "GET /rest/tourns/{tournId}/events/{eventAbbr} is undocumented. Need to add .openapi to handler",
        "parameters": [
          {
            "$ref": "#/components/parameters/tournId"
          },
          {
            "name": "eventAbbr",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/rest/tourns/{tournId}/invite": {
      "get": {
        "summary": "Get Tournament Invite",
        "operationId": "getTournInvite",
        "description": "Retrieve a public invite for a specific tournament, including pages, files, events, and contacts.",
        "tags": [
          "Tournaments",
          "test"
        ],
        "responses": {
          "200": {
            "description": "Public facing page data for a given tournament",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournInvite"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/tournId"
          }
        ]
      }
    },
...
    "/rest/tourns/{tournId}/results": {
      "get": {
        "summary": "Returns an array of result_sets that are published in a tournament",
        "operationId": "getTournPublishedResults",
        "tags": [
          "invite",
          "public",
          "results"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "tournId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "$ref": "#/components/parameters/tournId"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "description": "GET /rest/tourns/{tournId}/results is undocumented. Need to add .openapi to handler"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Category": {
        "type": "object",
        "required": [
          "id",
          "name",
          "abbr",
          "tournId"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "abbr": {
            "type": "string"
          },
          "tournId": {
            "type": "integer"
          },
          "pattern": {
            "type": "integer",
            "nullable": true
          },
          "settings": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "Events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Event"
            }
          },
          "Judges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Event"
            }
          }
        },
        "examples": {
          "example1": {
            "id": 1,
            "name": "Lincoln-Douglas",
            "abbr": "LD",
            "tournId": 42,
            "patternId": 7,
            "settings": [],
            "createdAt": "2023-01-01T00:00:00Z",
            "updatedAt": "2023-01-02T00:00:00Z"
          }
        }
      },
...
      "Contact": {
        "type": "object",
        "description": "A tournament contact for a school",
        "required": [
          "id",
          "schoolId",
          "personId"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "example": 42
          },
          "schoolId": {
            "type": "integer"
          },
          "personId": {
            "type": "integer"
          },
          "official": {
            "type": "boolean",
            "description": "True if the contact is a designated official contact for the school"
          },
          "onsite": {
            "type": "boolean",
            "description": "True if the contact will be physically present at the tournament"
          },
          "email": {
            "type": "boolean",
            "description": "True if the contact should get mass emails for the school"
          },
          "book": {
            "type": "boolean",
            "description": "True if the contact should appear in the Nationals TBook"
          },
          "nsda": {
            "type": "integer",
            "description": "NSDA ID number of the contact if they are a coach"
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "middleName": {
            "type": "string",
            "nullable": true,
            "example": "Quincy"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "state": {
            "type": "string",
            "example": "CA"
          },
          "country": {
            "type": "string",
            "example": "USA"
          },
          "tz": {
            "type": "string",
            "example": "America/Los_Angeles"
          },
          "createdAt": {
            "type": "string",
            "readOnly": true,
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "settings": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "Person": {
            "$ref": "#/components/schemas/Person"
          },
          "School": {
            "$ref": "#/components/schemas/School"
          }
        }
      },
      "Event": {
        "type": "object",
        "description": "An event",
        "required": [
          "id",
          "name",
          "abbr",
          "type",
          "categoryId",
          "tournId"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "abbr": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "fee": {
            "type": "number"
          },
          "type": {
            "type": "string",
            "enum": [
              "debate",
              "speech",
              "mockTrial",
              "congress",
              "wsdc",
              "wudc",
              "attendee",
              "academic"
            ]
          },
          "categoryId": {
            "type": "integer"
          },
          "settings": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "Tourn": {
            "$ref": "#/components/schemas/Category"
          },
          "Category": {
            "$ref": "#/components/schemas/Category"
          },
          "Topic": {
            "$ref": "#/components/schemas/Topic"
          },
          "Entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Entry"
            }
          },
          "nsdaCategoryId": {
            "type": "integer",
            "nullable": true
          },
          "NSDACategory": {
            "$ref": "#/components/schemas/NSDACategory"
          }
        }
      },
      "File": {
        "type": "object",
        "description": "A file",
        "properties": {
          "id": {
            "type": "integer",
            "description": "The unique identifier for the file",
            "readOnly": true,
            "example": 123
          },
          "tag": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "label": {
            "type": "string",
            "nullable": true
          },
          "filename": {
            "type": "string",
            "maxLength": 255,
            "description": "The filename of the file",
            "example": "example.pdf"
          },
          "published": {
            "type": "boolean",
            "description": "Whether the file is published or not",
            "example": true
          },
          "pageOrder": {
            "type": "integer",
            "format": "int16",
            "nullable": true
          },
          "uploaded": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the file was last modified",
            "readOnly": true,
            "example": "2023-10-26T15:51:32.000Z"
          }
        }
      },
...
      "TournInvite": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Tourn"
          },
          {
            "type": "object",
            "properties": {
              "Webpages": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Webpage"
                }
              },
              "Files": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/File"
                }
              },
              "Events": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EventInvite"
                }
              },
              "Contacts": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TournContact"
                }
              }
            }
          }
        ]
      },
...
      "NSDACategory": {
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "code"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "code": {
            "type": "integer"
          },
          "nationals": {
            "type": "boolean"
          }
        },
        "examples": {
          "example1": {
            "id": 202,
            "name": "Extemporaneous",
            "type": "s",
            "code": 202
          }
        }
      },
      "Problem": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "description": "A URI reference that identifies the problem type."
          },
          "title": {
            "type": "string",
            "description": "A short, human-readable summary of the problem type."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "The HTTP status code."
          },
          "detail": {
            "type": "string",
            "description": "Human-readable explanation of the error."
          },
          "instance": {
            "type": "string",
            "format": "uri",
            "description": "A URI reference to the specific occurrence."
          }
        }
      },
...
      "Tourn": {
        "type": "object",
        "description": "A tournament",
        "required": [
          "id",
          "name",
          "tz",
          "webname",
          "start",
          "end"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "The unique identifier for the tournament",
            "readOnly": true,
            "example": 123
          },
          "name": {
            "type": "string",
            "description": "The name of the tournament",
            "example": "National Speech & Debate Tournament"
          },
          "city": {
            "type": "string",
            "description": "The city where the tournament is held",
            "example": "Phoenix/Mesa"
          },
          "state": {
            "type": "string",
            "description": "The state where the tournament is held",
            "example": "AZ"
          },
          "country": {
            "type": "string",
            "description": "The country where the tournament is held",
            "example": "US"
          },
          "tz": {
            "type": "string",
            "description": "The IANA timezone of the tournament location",
            "example": "America/Chicago"
          },
          "webname": {
            "type": "string",
            "description": "The web name of the tournament",
            "example": "nationals"
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "description": "The start date and time of the tournament",
            "example": "2023-06-11T15:00:00.000Z"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "description": "The end date and time of the tournament",
            "example": "2023-06-17T05:00:00.000Z"
          },
          "regStart": {
            "type": "string",
            "format": "date-time",
            "description": "The registration start date and time of the tournament",
            "example": "2023-03-15T13:00:00.000Z"
          },
          "regEnd": {
            "type": "string",
            "format": "date-time",
            "description": "The registration end date and time of the tournament",
            "example": "2023-06-02T06:59:00.000Z"
          },
          "Contacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Contact"
            }
          },
          "Categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Category"
            }
          },
          "Events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Event"
            }
          },
          "Schools": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/School"
            }
          },
          "Webpages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Webpage"
            }
          },
          "Files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/File"
            }
          },
          "Emails": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Email"
            }
          },
          "createdAt": {
            "type": "string",
            "readOnly": true,
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "settings": {
            "type": "object",
            "additionalProperties": {
              "type": [
                "string",
                "integer",
                "boolean"
              ]
            }
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": [
                "string",
                "integer",
                "boolean"
              ]
            }
          }
        }
      },
      "TournContact": {
        "type": "object",
        "description": "A tournament contact person",
        "properties": {
          "id": {
            "type": "integer",
            "description": "The unique identifier for the contact person",
            "readOnly": true,
            "example": 456
          },
          "first": {
            "type": "string",
            "description": "The first name of the contact person",
            "example": "John"
          },
          "middle": {
            "type": "string",
            "nullable": true,
            "description": "The middle name of the contact person",
            "example": "A."
          },
          "last": {
            "type": "string",
            "description": "The last name of the contact person",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "The email address of the contact person",
            "example": "johndoe@example.com"
          }
        }
      },
...
      "Webpage": {
        "type": "object",
        "description": "A webpage",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "description": "Unique identifier",
            "readOnly": true
          },
          "title": {
            "type": "string",
            "maxLength": 63,
            "nullable": true,
            "description": "Title of the page"
          },
          "content": {
            "type": "string",
            "nullable": true,
            "description": "Page content"
          },
          "published": {
            "type": "boolean",
            "description": "Whether the page is published"
          },
          "sitewide": {
            "type": "boolean",
            "description": "Whether the page is visible sitewide"
          },
          "special": {
            "type": "string",
            "maxLength": 15,
            "nullable": true,
            "description": "Special page identifier or mode"
          },
          "slug": {
            "type": "string",
            "maxLength": 63,
            "nullable": true,
            "description": "page slug"
          },
          "pageOrder": {
            "type": "integer",
            "format": "int16",
            "nullable": true,
            "description": "Ordering value for page display"
          },
          "parentId": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "description": "Parent page ID (for hierarchical pages)"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the page was last modified",
            "readOnly": true,
            "example": "2023-10-26T15:51:32.000Z"
          }
        }
      },
...
      "School": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true,
            "description": "Unique identifier for the school"
          },
          "name": {
            "type": "string",
            "description": "Name of the school"
          },
          "code": {
            "type": "string",
            "description": "School code"
          },
          "onsite": {
            "type": "boolean",
            "description": "Whether the school is onsite"
          },
          "tournId": {
            "type": "integer",
            "description": "Tournament ID"
          },
          "chapterId": {
            "type": "integer",
            "description": "Chapter ID"
          },
          "state": {
            "type": "string",
            "description": "State abbreviation"
          },
          "regionId": {
            "type": "integer",
            "description": "Region ID"
          },
          "districtId": {
            "type": "integer",
            "description": "District ID"
          },
          "updatedAt": {
            "type": "string",
            "readOnly": true,
            "format": "date-time",
            "description": "Last modified timestamp"
          },
          "createdAt": {
            "type": "string",
            "readOnly": true,
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "Chapter": {
            "$ref": "#/components/schemas/Chapter"
          },
          "Entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Entry"
            }
          },
          "Judges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Judge"
            }
          },
          "settings": {
            "type": "object",
            "additionalProperties": {
              "type": [
                "string",
                "integer",
                "boolean"
              ]
            }
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": [
                "string",
                "integer",
                "boolean"
              ]
            }
          }
        },
        "additionalProperties": false,
        "examples": {
          "example1": {
            "id": 1985,
            "name": "Shermer High School",
            "code": "SHS",
            "onsite": true,
            "tournId": 123,
            "chapterId": 456,
            "state": "IL",
            "regionId": 789,
            "districtId": 987,
            "settings": {
              "allowGuests": true
            },
            "updatedAt": "2024-01-15T12:34:56Z",
            "createdAt": "2023-09-01T08:00:00Z"
          }
        }
      },
...
    },
    "responses": {
      "ErrorResponse": {
        "description": "Unexpected Error",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            },
            "examples": {
              "error": {
                "summary": "500 Internal Server Error",
                "value": {
                  "type": "about:blank",
                  "title": "Internal Server Error",
                  "status": 500,
                  "detail": "An unexpected error occurred.",
                  "instance": "/api/v1/example"
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "NotFound - the requested resource was not found or you do not have access.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            },
            "examples": {
              "notfound": {
                "summary": "404 Not Found",
                "value": {
                  "type": "about:blank",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "The requested resource foo with ID bar was not found.",
                  "instance": "/api/v1/example"
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized - authentication failed or was not provided.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            },
            "examples": {
              "unauthorized": {
                "summary": "401 Unauthorized",
                "value": {
                  "type": "about:blank",
                  "title": "Unauthorized",
                  "status": 401,
                  "detail": "You are not authorized to access this resource."
                }
              }
            }
          }
        }
      }
    },
    "parameters": {
      "tournId": {
        "name": "tournId",
        "in": "path",
        "description": "The ID or webname of the tournament to retrieve.",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "securitySchemes": {
      "extApiKey": {
        "type": "http",
        "scheme": "basic"
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "x-tabroom-cookie"
      }
    }
  }
}